requirements.txt
requests~=2.28.2
pytest~=7.2.2
voluptuous~=0.13.1
pytest-voluptuous
selene==2.0.0b16
python-dotenv==1.0.0
allure-pytest==2.12.0
curlify~=2.2.1
selenium~=4.4.3
flaky~=3.7.0